some mistake

+++ signup.js
@@ -1,6 +1,8 @@
 let user = JSON.parse(localStorage.getItem("user")) || {};
 
 function signup(e) {
+
+  
   e.preventDefault();
   let username = document.getElementById("name");
   let useremail = document.getElementById("email");
@@ -12,9 +14,24 @@ function signup(e) {
     password: userpassword.value
   };
 
-  localStorage.setItem("user", JSON.stringify(user));
+
+  if (user.name == '' || user.email == '' || user.password == '') {
+   Swal.fire({
+    icon: 'error',
+    title: 'Oops...',
+    text: 'Please fill in all fields'
+  });
+  }else{
+    localStorage.setItem("user", JSON.stringify(user));
+   Swal.fire({
+    icon: 'success',
+    title: 'Success!',
+    text: 'Sign up successful',
+    confirmButtonText: 'OK'
+  }).then(() => {
+    window.location.href = "login.html";
+  });
+  }
 
 
-  window.location.href = "login.html";
 }
-
